+++ mods/CreateDragonPlus/src/main/java/plus/dragons/createdragonsplus/common/fluids/hatch/FluidHatchBlock.java
@@ -214,8 +214,11 @@ public class FluidHatchBlock extends HorizontalDirectionalBlock implements IBE<F
             fluidCopy.setAmount(requiredAmountForItem);
             capability.drain(fluidCopy, FluidAction.EXECUTE);
 
-            if (!player.isCreative())
-                player.getInventory().placeItemBackInInventory(result);
+            if (!player.isCreative()){
+                if(player.getItemInHand(hand).isEmpty())
+                    player.setItemInHand(hand,result);
+                else player.getInventory().placeItemBackInInventory(result);
+            }
             blockEntity.setChanged();
             if (level instanceof ServerLevel serverLevel)
                 serverLevel.getChunkSource().blockChanged(blockEntity.getBlockPos());