+++ mods/CreateDragonPlus/CHANGELOG.md
@@ -1,11 +1,14 @@
 ### Create: Dragons Plus 1.7.0
 
-#### Updaye
+#### Update
 * Add Hungarian Translation (by @Batári Balázs László)
 * Add block tag for automatic polishing recipe generation denylist
 * Add Stainable attribute for the attribute filter
 * Bucket of Mod Fluids works properly with Dispenser
 * Liquid Dye now can solidify concrete powder
+
+#### Change
+* Keep Bucket on Main Hand after filling from Fluid Hatch
 
 #### Fix
 * Fix fan processing is still running when corresponding recipe is disabled in config.